spli cli messages to new file
