+++ src/extension/src/constants/cli.ts
@@ -1,0 +1,26 @@
+
+//reagrupar por request/result
+export const CLI = {
+  BASE_CLI_TOOL_NAME: "Microsoft.Templates.Cli",
+  SYNC_COMPLETE_STATE: "syncResult",
+  SYNC_PROGRESS_STATE: "syncProgress",
+  GET_FEATURES_COMPLETE_STATE: "getFeaturesResult",
+  GET_FRAMEWORKS_COMPLETE_STATE: "getFrameworksResult",
+  GET_PAGES_COMPLETE_STATE: "getPagesResult",
+  GET_PROJECT_TYPES_COMPLETE_STATE: "getProjectTypesResult",
+  GENERATE_COMPLETE_STATE: "generateResult",
+  GENERATE_PROGRESS_STATE: "generateProgress",
+  WINDOWS_PLATFORM_VERSION: "win32",
+  SYNC_COMMAND_PREFIX: "sync",
+  GET_FRAMEWORKS_COMMAND_PREFIX: "getframeworks",
+  GET_ALL_LICENSES_COMMAND_PREFIX: "getalllicences",
+  GET_ALL_LICENSES_COMPLETE_STATE: "getAllLicencesResult",
+  GET_PAGES_COMMAND_PREFIX: "getpages",
+  GET_FEATURES_COMMAND_PREFIX: "getfeatures",
+  GET_PROJECT_TYPES_COMMAND_PREFIX: "getprojecttypes",
+  GENERATE_COMMAND_PREFIX: "generate",
+  PRODUCTION_PATH_TO_TEMPLATES: "..",
+  DEVELOPMENT_PATH_TO_TEMPLATES: "../../../../..",
+  MAX_SYNC_REQUEST_ATTEMPTS: 51,
+  SYNC_RETRY_WAIT_TIME: 250,
+};
