+++ src/extension/src/utils/validator.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { CLI } from "../constants/extension";
+import { CLI } from "../constants/cli";
 import { ExtensionCommand } from "../constants/extension";
 import fs = require("fs");
 import path = require("path");
